pyproject.toml
[tool.poetry]
name = "dreamhouse-db"
version = "0.1.0"
description = ""
authors = ["Gamaliel Garcia <46827955+EGAMAGZ@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dreamhouse_db"}]

[tool.poetry.dependencies]
python = "^3.11"
faker = "^18.10.1"
mysql-connector-python = "^8.0.33"
simple-query-builder = "^0.3.6"
mariadb = "^1.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["W", "F", "I", "N", "ANN"]
fixable = ["I", "F", "ANN"]